Move Java toolchain integration tests to appropriate package
